src/img-gen/gemini/requirements.txt
git+https://github.com/google-gemini/generative-ai-python@imagen
python-dotenv
Pillow
